feat: sitter galleries and profile upload
